+++ src/main.rs
@@ -83,6 +83,7 @@ mod util;
 mod windows;
 mod worker;
 
+mod completions;
 #[cfg(test)]
 mod tests;
 
@@ -92,7 +93,6 @@ use crate::{
         ChatStore,
         HomeserverAction,
         IambAction,
-        IambCompleter,
         IambError,
         IambId,
         IambInfo,
@@ -102,6 +102,7 @@ use crate::{
         ProgramContext,
         ProgramStore,
     },
+    completions::IambCompleter,
     config::{ApplicationSettings, Iamb},
     windows::IambWindow,
     worker::{create_room, ClientWorker, LoginStyle, Requester},
